Restructuring and renaming for migration to forseti-security repo
