+++ rpe/exceptions.py
