+++ rpe/resources/__init__.pyc
