+++ tox.ini
@@ -7,4 +7,4 @@ deps =
        google-api-python-client-helpers
        pytest
        pytest-cov
-commands = pytest --cov=micromanager --capture=no --cov-report=term-missing
+commands = pytest --cov=rpe --capture=no --cov-report=term-missing
